Fixing compiler warnings in SNMP file.
There was a variable declared that was only used if building for 6100 or 6300 chips. Changed that to use the same #if that is used to include that code block, removing the warning. There was also a scanf() using %u for uint32_t variables. On systems where % is not 32-bits, this generates a compiler warning. Including inttypes.h gives access to PRIu32 which will be "whatever printf format represents uint32_t" on the compiler. If an old or lacking compiler is used without inttypes.h, the original line is still in there commented out. Maybe it needs a comment to explain this?